Process tickets limits
Use the limit to enforce tickets that should be purchased individually,
and use float reminder to acts as a probability or purchasing an extra
ticket.

+++ config.go
@@ -166,8 +166,8 @@ type ticketBuyerOptions struct {
 	BalanceToMaintainAbsolute *cfgutil.AmountFlag  `long:"balancetomaintainabsolute" description:"Amount of funds to keep in wallet when purchasing tickets"`
 	VotingAddress             *cfgutil.AddressFlag `long:"votingaddress" description:"Purchase tickets with voting rights assigned to this address"`
 	votingAddress             dcrutil.Address
-	Limit                     uint   `long:"limit" description:"Buy no more than specified number of tickets per block (0 disables limit)"`
-	VotingAccount             string `long:"votingaccount" description:"Account used to derive addresses specifying voting rights"`
+	Limit                     float64 `long:"limit" description:"Buy no more than specified number of tickets per block (0 disables limit)"`
+	VotingAccount             string  `long:"votingaccount" description:"Account used to derive addresses specifying voting rights"`
 }
 
 // cleanAndExpandPath expands environement variables and leading ~ in the
